# PERF: more flexible iso8601 parsing