--- a/asv_bench/benchmarks/timeseries.py
+++ b/asv_bench/benchmarks/timeseries.py
@@ -1059,32 +1059,26 @@ class timeseries_to_datetime_iso8601(object):
     goal_time = 0.2
 
     def setup(self):
-        self.N = 100000
-        self.rng = date_range(start='1/1/2000', periods=self.N, freq='T')
-        if hasattr(Series, 'convert'):
-            Series.resample = Series.convert
-        self.ts = Series(np.random.randn(self.N), index=self.rng)
         self.rng = date_range(start='1/1/2000', periods=20000, freq='H')
         self.strings = [x.strftime('%Y-%m-%d %H:%M:%S') for x in self.rng]
+        self.strings_nosep = [x.strftime('%Y%m%d %H:%M:%S') for x in self.rng]
+        self.strings_tz_space = [x.strftime('%Y-%m-%d %H:%M:%S') + ' -0800'
+                                 for x in self.rng]
 
     def time_timeseries_to_datetime_iso8601(self):
         to_datetime(self.strings)
 
-
-class timeseries_to_datetime_iso8601_format(object):
-    goal_time = 0.2
-
-    def setup(self):
-        self.N = 100000
-        self.rng = date_range(start='1/1/2000', periods=self.N, freq='T')
-        if hasattr(Series, 'convert'):
-            Series.resample = Series.convert
-        self.ts = Series(np.random.randn(self.N), index=self.rng)
-        self.rng = date_range(start='1/1/2000', periods=20000, freq='H')
-        self.strings = [x.strftime('%Y-%m-%d %H:%M:%S') for x in self.rng]
+    def time_timeseries_to_datetime_iso8601_nosep(self):
+        to_datetime(self.strings_nosep)
 
     def time_timeseries_to_datetime_iso8601_format(self):
         to_datetime(self.strings, format='%Y-%m-%d %H:%M:%S')
+
+    def time_timeseries_to_datetime_iso8601_format_no_sep(self):
+        to_datetime(self.strings_nosep, format='%Y%m%d %H:%M:%S')
+
+    def time_timeseries_to_datetime_iso8601_tz_spaceformat(self):
+        to_datetime(self.strings_tz_space)
 
 
 class timeseries_with_format_no_exact(object):