@jorisvandenbossche 2016-01-16
A question: is is worth the added complexity to the ISO 8601 C-parser to be able to handle the dates without separator (which is actually not really ISO 8601 anymore), if when passing the `format` is already as fast to parse this? 
Of course, you then need to specify this, but this could also be alleviated by making `infer_datetime_format=True` the default.