@chris-b1 2016-01-16
Hard to say if it's worth it (re dates with separators), not sure it's a super common format.  But there is a substantial speedup even over providing a format to the strptime parser (20x in that asv bench).  

Not that it really matters, but from I've read dates without separators would be considered ISO 8601, e.g. http://www.cl.cam.ac.uk/~mgk25/iso-time.html.