@jreback 2016-01-16
how hard would it to allow a selection of separators

eg -, ., /, \ (assume that if one is found it would have to repeat for the date portions, iow don't allow mixed seps)

w/o sacrificing ambiguity?