@jreback 2016-01-16
also, let's add another issue to make ``infer_datetime_format=True`` as the default for ``to_datetime`` and ``read_csv``. I don't think this will impact anything and just make more things work.