@chris-b1 2016-01-16
@jreback - I don't think it would be too bad to parse multiple separators, I'll try it out