@jreback 2016-01-19
@chris-b1 just some minor comments on the benchmarks themselves here.

if the addtl separators is a bit complicated can do in another PR.

pls rebase on master as lots of PEP changes recently and do a ``git diff master | flake8 --diff`` check

ping if you want to merge now (else, pls open a new issue for the other sep chars if not)