@chris-b1 2016-01-20
@jreback - updated for that asv note.  I was most of the way there already, so also updated to handle 
 1. date separator can be any of `{'-', '.', '/', '\\', ' '}`, as long as consistent
 2. leading 0s can also be omitted from the time components (must have `:` sep)

Passes the current test suite (adjusted for a couple timezone changes), not sure if there are other edge cases to try?

a couple perf regressions to figure out too

```
    before     after       ratio
  [567bc5ce] [82c5d7b9]
+  154.12ms      4.24s     27.50  plotting.plot_timeseries_period.time_plot_timeseries_period
+    1.82ms      2.19s   1203.52  timeseries.datetimeindex_infer_dst.time_datetimeindex_infer_dst
-     2.10s     5.14ms      0.00  timeseries.timeseries_to_datetime_iso8601.time_timeseries_to_datetime_iso8601_nosep
-     3.13s   223.79ms      0.07  timeseries.timeseries_to_datetime_iso8601.time_timeseries_to_datetime_iso8601_tz_spaceformat
```