@chris-b1 2016-01-21
@jreback  - if you want to have a look, I've got everything working now.