@jreback 2016-01-21
can you add tests for differnt seps (and invalid ones)?