@jreback 2016-01-22
did you figure out the perf regressions?