@chris-b1 2016-01-22
Well, I seemed to fix it, but didn't figure it out.  

`tz_localize` was very running very slow, but only a fresh install of pandas - wasn't slow if run a second time.  Also seemed to be fine if stepping through with `pdb`.  Undoing this [change](https://github.com/chris-b1/pandas/commit/5d565297cc653c80d0c3eaece864d457135c6876#diff-05de2950b2c86c9828531957e02d1b87L2021) fixed it - very unclear to me how that caused it though.