@jreback 2016-01-22
hmm, I think this interferes with the ``infer_datetime_format`` ? haven't really stepped thru the code that recently.